refactor(bin-designer): split types/index.ts into domain modules
types/index.ts was a 1343-line barrel that also DEFINED most of the
designer's types inline, mixing twelve domains and hiding runtime values
(style guards, const arrays) among interfaces. The inline content now
lives in one module per domain - base, dividers, compartments, labelTabs,
handles, walls (incl. overhang + pattern), splitConnector, binParams
(the 122-line mega-interface + Insert), generation (status/state/mesh/
history), uiState (DesignerUIState, tabs, export naming), savedDesign,
designerState - alongside the pre-existing cutout/featureColors/lid/
text/exampleGallery modules.

index.ts keeps its external imports and existing re-export blocks and
re-exports the new modules, so all 218 barrel importers are untouched.
Types moved verbatim; max-lines warning cleared.
